Streamlined the parallel tridiagonal solver
This commit includes various changes to make the parallel tridiagonal solver simpler
and more efficient:
- Several work arrays that depend only on the matrix (which is held fixed during the
  linear solve) and not the rhs vector (which varies from one iteration to the next)
  are now precomputed before the iterative solve, so that the number of operations
  per iteration is smaller.
- I restructured some arrays so that the same code used to solve the u component over
  grid rows, with arrays of size (ilocal,jlocal), can be used to solve the v component
  over grid columns, with arrays of size (jlocal,ilocal).  The horizontal indices are
  reversed for the v component to reduce strides.

This commit includes three versions of the parallel tridiagonal solver: (1) the solver from
the previous commit, (2) a newer version with separate code for u and v components,
and (3) the newest version, with code that applies to either the u or v component.
I checked that (2) and (3) give the same results, BFB.  Results from (1) differ at roundoff level.

A near-future commit will clean up the code further by removing (1) and (2).