Added a 2D matvec multiply subroutine with reordered loops
The matrix-vector multiplication subroutine in the linear solver accounts
for a large fraction of the cost of most CISM simulations.
The loops are ordered with the 'm' index first, followed by i and j.
The code will likely vectorize more efficiently if the order is (i,j,m).

This commit adds an optional new subroutine, matvel_multiply_structured_2d_new,
that works with matrices having the (i,j,m) ordering.  On a Macbook,
the cost is not much changed, but there may be larger differences on Cheyenne
using the intel compiler with vector directives. This is to be tested.

I compared answers with the old and new subroutines.
For the 4-processor dome DIVA problem and for standard LIVV tests,
results are BFB, which was unexpected given that the order of operations
is different.  I am not sure results would be BFB on other machines or problems.

For now, the (m,i,j) ordering remains the default.