Commit reversion + more bugfixes Reverting changes from the physics/discretization commit 0b33606, such that damage is no longer converted to crevasse depth during the course of the damage evolution routine.
Also, this commit extends the unstaggered velocity extrapolation procedure to occur at the grounding line in addition to the calving front, where there are not enough staggered grid points for the velocity to be defined. Some of this procedural extension involves minor bugfixes in identifying the correct points within the domain.