Reduced cost of global gather/scatters for tridiag preconditioning
For the confined-shelf and Antarctic problems, tridiagonal preconditioning reduces
the number of linear iterations by a factor of about 2 compare to diagonal preconditioning
with the same convergence threshold.  This results in a cost savings, but some or all
of the savings is lost by doing gathers and scatters to solve a global tridiagonal problem
for each row and column of the grid.

This commit reduces the cost of global gathers and scatters.  I tried several changes:
(1) Call fc_gather_int (to compute an array of indices of task boundaries) only once,
    instead of every time row and column data are gathered or scattered.
    This gives a significant savings and is BFB.
(2) Reduce the maximum number of iterations from 200 to 100 when doing tridiagonal
    preconditioning.  As a rule of thumb (at least for Antarctica), tridiagonal PC
    reaches about the same convergence threshold in half the number of linear iterations as
    diagonal PC, so this change results in about the same number of convergence failures
    in the first several linear iterations.  From experience, convergence failures are not a
    terrible thing; the solver usually gets pretty close to the solution.
    This parameter change is answer-changing for tridiagonal PC.
(3) Tried changing the tridiagonal solution algorithm from
    (a) gather global data on main task, solve on main task, and scatter the solution back to local tasks,
    to
    (b) gather global data on each task and solve the global problem on each task; no scatter needed.
    This change is BFB.  It did not significantly speed up the code in any tests I tried.
    I left the option in the code; it can be switched on by setting gather_all = .true.
    in subroutine pcg_solver_chrongear_2d.

To implement (3), I added two new subroutines in parallel_mpi.F90.
These subroutines use mpi_allgather and mpi_allgatherv to gather data to all tasks
in a row or column.  I'm not aware of an allgather anaolog to the fc_gather routines,
but it's possible that such a routine could reduce costs for the gather_all option.

I also turned off some code in glissade.F90 that was generating test arrays
for row and column communications at initialization.

With these changes, the parallel tridiagonal preconditioner seems ready for production.
So far it has been implemented only for the 2D Chronopoulos-Gear PCG solver.
At some point we can add this option to the 2D standard PCG solver.