Switched matrix index order for the 2D velocity solver
Until now, the assembled global matrices for the 2D velocity solver have
had dimensions (nNodeNeighbors_2d,nx-1,ny-1), where nNodeNeighbors = 9
includes the cell and its nearest neighbors in the horizontal direction.

This commit switches the order to (nx-1,ny-1,nNodeNeighbors).
The goal is to make the code more vector-friendly, especially for matrix-vector
multiplies and halo updates, by putting the smallest dimension last.

I verified that for the dome and shelf problems on a Mac, this commit is BFB,
even though the order of sums in the matvec multiply has changed.
For the dome, I checked BFB for both the Chronopoulos-Gear and standard 2D solvers.
However, results may not be BFB for other problems and platforms.

Still to do is to switch the order of indices for the 3D Blatter-Pattyn solve.