Added linear_solve_ncheck as a config option
I added a new user-configurable option, linear_solve_ncheck.
This is the number of iterations between convergence checks in the Glissade PCG solvers.
The goal is to choose a value that is not too frequent (because each convergence check
requires a matvec multiply and global sum), but also not too infrequent (so as to avoid
doing too many additional iterations after the linear solver has converged).

Until now, this parameter was called solve_ncheck and was set to a hardwired value of 5.
The default remains 5 (ensuring BFB results).
For problems with many linear iterations (~100 or more for the first few nonlinear iterations),
a larger value (perhaps between 5 and 10) might reduce the solver cost by a few percent.

One exception to the general rule: A convergence check is always on the 5th iteration,
even if linear_solve_ncheck > 5, because the last few nonlinear iterations typically
converge in 5 linear iterations or less.

Other minor cleanup:
- Set verbose_pcg in glissade_velo_higher_pcg.F90 instead of passing it as an input argument
- Made itest, jtest, and rtest required (instead of optional) arguments for the PCG solver subroutines
- Changed the linear iteration counter from 'n' to 'iter' for easier grepping.

This commit is BFB except when linear_solver_ncheck is set to a value other than 5.