Created separate local and global tridiagonal preconditioner options
This commit divides the new tridiagonal preconditioning option into two options,
local and global:
  integer, parameter :: HO_PRECOND_TRIDIAG_LOCAL  = 3
  integer, parameter :: HO_PRECOND_TRIDIAG_GLOBAL = 4

The global option preconditions by doing a full global solve of the system T*x = b,
where T is a tridiagonal approximation of the full matrix (specifically, of the
submatrices Auu and Avv).  This requires gathering data onto the main task in each
row or column of the grid to solve a tridiagonal system for that row or column.

I put a lot of effort into making the global gather more efficient, including
(in this commit) reducing the size of the buffers to be gathered.
But I've been unable (for an 8km Antarctic problem on Cheyenne) to make the global
tridiagonal solver efficient enough for production.
The savings from doing fewer linear iterations (hence fewer matvec multiplies and
global sums) are entirely canceled by the cost of the preconditioning, which is
dominated by the cost of gather/scatters.

On the bright side, it turns out that local block tridiagonal preconditioning
(i.e., doing a tridiagonal solve on each task, without any global communication)
significant reduces the number of linear iterations.  The number of iterations is more
than for the global tridiagonal solve, but much less than for diagonal preconditioning.
Since there is little overhead for communication, the local tridiagonal option gives
significant speedup.

It might be possible to speed up the local option further by eliminating halo updates;
this remains to be tested.

Related changes:

- I added two new subroutines to handle option HO_PRECOND_TRIDIAG_LOCAL, and removed
  some subroutines that are no longer used.
  Note: The local solver does the Auu and Avv solves in sequence in a single subroutine call.
        In contrast, the global solver is called twice, once for Auu and once for Avv.
        This requires some rearrangement of arrays before and after the call, but has
        the advantage of avoiding duplication of a long, complex piece of code,
        and also reducing strides for the column solve.

- I modified the row- and column-based gathers and scatters in parallel_mpi.F90
  to reuse the saved value of d_gs_bounds_row or d_gs_bounds_column, unless
  the input buffer size has changed since the previous call.  This reduces
  some unnecessary gathers without hardwiring d_gs_bounds for the entire run.

- I switched the default of verbose_solver to T in glissade_velo_higher.F90,
  since I've invariably been turning on this output when looking at code speedup.

- I tested the following logic change in the PCG solvers:
     from this:   if (mod(iter, linear_solve_ncheck) == 0 .or. iter == 5) then
     to this:     if (mod(iter, linear_solve_ncheck) == 0 .or. iter == linear_solve_ncheck/2) then
  And I found some code speedup for linear_solve_ncheck > 5.  However, I kept the old version
  in the code for now, because the new version is answer-changing for nearly all runs.