Added option to enter smb_anomaly in units of mm/yr w.e.
In general, CISM can read in SMB forcing in units of m/yr ice (as acab) or
in units of mm/yr water equivalent (as smb), depending on the value of smb_input.

With this commit, smb_anomaly can similarly be part of an input file.
Previously, CISM could read in only acab_anomaly.

Also fixed a minor bug in logic for applying acab_anomaly_timescale.
The new logic avoids divzero's in first year of run when timescale = 0,
and avoids rounding errors when the time is close to an integer value.