Some changes that make sure array bounds are followed.
pass unphys_val down to interpolate_thermal_forcing_to_lsrf so can
check if thermal_forcing is set, if it isn't set it to zero.

change loops that start at nhalo to nhalo+1, so subscripts
remain within arrays.

Set indices for nearest neighbors so can ensure within array bounds.
Add more checks that both kbot and ktop are set to signify a valid
ocean point. If ktop == kbot set both to zero. If checks on kbot or
ktop, make sure it's not zero before doing the check.

Comment out the bug check that all points are filled. It writes
the error for several points. But, doing this allows the simulation
to complete.