Added a 2D scalar PCG solver with a test problem
This commit includes three new subroutines in module glissade_velo_higher_pcg.F90:
- pcg_solver_standard_2d_scalar
  This subroutine is based on pcg_solver_standard_2d, but instead of solving for horizontal
  velocity components (xu, xv) given matrices (Auu, Auv, Avu, Avv) and rhs arrays (bu, bv),
  it solves for a scalar xh given Ah and bh.
  The structure of the PCG algorithm is the same, but the code is simpler with fewer arrays.
- matvec_multiply_structured_2d_scalar
  This is based on matvec_multiply_structured_2d, but for the scalar problem.
- subroutine global_sum_scalar
  This is based on the global_sum_staggered subroutines, but for the scalar problem.

These subroutines are intended for use by the new basal hydrology scheme, which will
solve an elliptic equation for the hydrostatic head, a scalar that lives on the ice grid.

To test the new solver, I added a logical parameter 'test_matrix' in glissade.F90.
When test_matrix = .true., there is a call in glissade_initialise to a new subroutine
called glissade_test_matrix in glissade_test.F90.  This subroutine sets up a simple
matrix problem with 4's along the diagonals and with -1's along the 4 off-diagonals
corresponding to edge neighbors.  (This is similar to the stencil for the head solve.)
The rhs consists of 1's in cells adjacent to the global boundary and 0's elsewhere,
giving a simple solution of xh = 1 everywhere.

I verified that the new solver gives the right answer for the test problem, both in serial
(with 1 task) and in parallel (with 4 tasks).  The code writes out some diagnostics
and then aborts (as intended).

Later, it would be easy to write a Chronopoulos-Gear version of the scalar solver,
with one global sum per linear iteration instead of two to reduce the computational cost.

Since the new solver is not called from glissade_velo_higher.F90, the module name
glissade_velo_higher_pcg is no longer entirely accurate.  We might want to change
the name to something more generic like glissade_solver_pcg.

This commit is BFB for the default value, test_matrix = .false.