Added option to invert for deltaT_basin
This commit implements a new option, which_ho_dtocn_inversion.
As with inversion of Cp and bmlt_float, the choices are (0) no inversion,
(1) compute values by inversion, and (2) apply values from previous inversion.
(In the previous commit, this option was going to be called which_ho_dtf_inversion,
but 'dtocn' better suggests warming and cooling of the ocean.)

The field for which we invert is deltaT_basin, a temperature bias-correction term
in the ISMIP6 nonlocal basal melt parameterizations (with or without slope weighting).
Although deltaT_basin is a 2D field, it is spatially uniform within each basin,
so there are far fewer degrees of freedom than for bmlt_float inversion.

Since there is only one DOF per basin, we need a basin-level inversion target.
At least for now, the inversion target is marine_grounded_area_target, which is
set to the initial (i.e., observed) area of grounded marine ice per basin.
Arguably, this is a better target than the floating ice area (which is sensitive
to the calving front location) and the total area of grounded ice (which can
include vast areas of ice grounded above sea level, disconnected from the ocean).

The algorithm is:
- At initialization, compute marine_grounded_area_target in each basin.
  Call this A_mgt.
- During each step at runtime, sum over all tasks to find the current value
  of marine_grounded_area in each basin.  Call this A_mg.
- Compute the ratio r = (A_mg - A_mgt) / A_mgt.
- Where r > 0, increase deltaT_basin to promote GL retreat.
  Where r < 0, decreate deltaT_basin to promote GL advance.
  The change is given by r * dtocn_dt_scale * dt, where dtocn_dt_scale is a new
  user-configurable parameter in the inversion derived type.

Notes:
- This inversion option is supported only for bmlt_float = 6 (BMLT_FLOAT_THERMAL_FORCING).
  The code aborts if the user selects a different value of bmlt_float.
- This option is much simpler to implement than bmlt_float inversion.
  The rate of change of deltaT_basin depends on only one parameter, dtocn_dt_scale.
- Although marine_grounded_area_target has only as many values as there are basins,
  it is defined as a 2D field for convenience in I/O (so it is not required to add nbasin
  as a dimension in netCDF files).
- For now I am using the 16 IMBIE basins defined for ISMIP6. But it might be helpful to
  have more basins and thus more DOFs (e.g., to separate PIG and Thwaites into two basins).
- For now I am reading in deltaT_basin as computed by Nico Jourdain for ISMIP6,
  but this is not necessary; the run could simply start from deltaT_basin = 0.

I have tested the new code in short runs on a Mac. Will next test in longer runs on Cheyenne.

This commit is BFB for runs that do not use the new which_ho_dtocn_inversion option.