Changed the computation of weight_ground_vertex
Earlier, weight_ground_vertex was defined in such a way as to reduce basal friction
at vertices adjacent to lightly grounded ice.
This commit uses a new definition, which has the effect of increasing basal friction
at vertices adjacent to shallow ocean cavities.  The idea is that if the cavity thickness
is comparable to a certain thickness scale (beta_cavity_thck_scale), then there are
likely to be basal protrusions that reach the bed, increasing the effective value
of f_ground.

When ocean cavities are thick compared to bmlt_cavity_thck_scale, then weight_ground_vertex
reduces to f_ground.

The motivation for this change is that there are regions where the ice is lightly grounded
(e.g., Thwaites Glacier near the calving front) and the observed surface velocity
shows evidence of grounding, but the model simulates little or no basal friction.
By setting beta_cavity_thck_scale > 0, it will be possible to increase the friction.

For now, weight_ground_vertex is not used, but it could be passed to calcbeta
by the velocity solver.

It is a bit confusing to have two different versions of f_ground at vertices.
In a future commit, I may combine these into a single variable.

Also removed an unnecessary argument, f_ground, from subroutine glissade_inversion_bmlt_basin.

This commit is BFB for the tests I ran.  However, there might be roundoff-level changes
associated with a change in quadrant indexing in the f_ground computation.