Fixed bug in temperature initialization with SMB input forcing
Surface mass balance can be read in with units of m/yr ice (acab, with smb_input = 0)
or mm/yr w.e. (smb, with smb_input = 1).
In the latter case, acab needs to be computed from the input smb before calling
glissade_init_therm, if the temperature is to be initialized correctly for
case init_temp = 3 (where advective-diffusive balance is assumed iff acab > 0).
Without this initialization, acab = 0 when glissade_init_therm is called,
leading to a too-warm column in many grid cells.

This commit adds the missing acab computation, fixing the initial temperature profiles.

Also, artm_corrected is computed at initialization, so that the initial value of artm
appears correctly in the log file.

This commit is answer-changing (fixing the bug) for runs with smb_input = 1.