Adding a new test case for Shakti
This commit adds a new test case called HYDRO in the 'tests' directory.
The README file contains information about the detail of the test.
Right now, the test sets up a linear bed, a uniform ice slab with a
constant sliding velocity.
This test is still under development and will be modified along with the
implementation of Shakti.