Last bmlt_basin changes before ISMIP6 runs for CISM Antarctic paper
A few tweaks to support production runs for ISMIP6 Antarctic experiments:

(1) One more change in the target for bmlt_basin inversion.  Previously, the target grid cells
had either (a) floating ice, or (b) lightly grounded ice, as defined by a threshold value
of thickness above flotation.
I decided to leave out floating ice with very deep cavities, because this ice is unlikely to ground.
Instead, we now compute f_flotation for each cell:
        f_flotation = (topg - eus) - (rhoi/rhoo)*thck
If positive, this quantity is the cavity thickness.
If negative, it is not a cavity thickness, but its magnitude indicates how close a grounded cell is to floating.
The new criterion include cells in the inversion target domain if abs(f_flotation) < bmlt_basin_cavity_threshold.
The threshold has a default value of 400 m, which seems to include most regions that
are likely to flip between grounded and floating.  This value can be set in the config file.

(2) I made a minor change to support option which_ho_bmlt_basin_inversion = 2, i.e., applying
deltaT_basin as computed from a previous inversion:
The user can now include thermal_forcing in the input file, and not have to specify thermal_forcing_baseline.
The code sets thermal_forcing = thermal_forcing_baseline only if nonzero thermal_forcing_baseline is read in.

(3) Added a max_bmlt diagnostic.  In glide_diagnostics.F90, we now identify the grid cell
with the largest basal melt rate, and write the rate and the global cell indices to the log file.

(4) When making the ice_domain_mask to support option compute_blocks = 1, extend the domain from
initially ice-covered cells 3 times instead of 2.  This captures some cells that, if not in the mask,
can cause the velocity solver to fail.