Minor change to prevent a rare divzero
In subroutine glissade_vertical_remap, there are checks that the total column thickness
('thck' and single-layer thickness 'hlyr' are greater than 0.
In rare cases (encountered during a 2-km Antarctic spin-up), we can have thck very slightly > 0,
but still get a divzero error when dividing by hlyr.

The fix is to require thck and hlyr to be greater than tiny(0.0d0).
This avoids the divzero for the run that crashed.