Added an option to adjust the input topography
This commit adds an option to adjust the input bed topography in a specified rectangular region.
The region is specified in terms of the max and min values of the x and y coordinates:
xmin, xmax, ymin, and ymax.

The adjustment is defined by three variables: topg_lo, topg_hi, and topg_delta.
- Where topg <= topg_lo, there is no adjustment.
- Where topg >= topg_hi, we apply the maximum adjustment, topg_delta.
- Where topg_lo < topg < topg_hi, the adjustment is phased in as a linear function of topg.

The adjustment is applied by setting adjust_input_topography = .true.
in the [options] setting of the config file.  This option is false by default.
Also, the user should choose values for 7 config parameters: adjust_topg_xmin, adjust_topg_xmax,
adjust_topg_ymin, adjust_topg_ymax, adjust_topg_lo, adjust_topg_hi, and adjust_topg_delta.
These are set in the [parameters] section of the config file.

To compute the adjustment, I wrote a new subroutine, glissade_adjust_topography.
This subroutine didn't fit obviously in an existing module, so I put it in a new file, glissade_utils.F90.
I then moved two other chunks of code from glissade.F90 to the new glissade_utils module,
in subroutines glissade_adjust_thickness and glissade_smooth_topography.
All three subroutines modify the input ice thicknesss and/or topography at initialization.
Other subroutines could be moved to glissade_utils.F90 in the future.

I tested the new option by raising the topography by 100 to 150 m near the calving front
in the eastern part of the Thwaites ice shelf.  This is a region where the ice is grounded in reality,
producing an ice rise, but the ice often does not ground in simulations.
The topography is not well constrained by data (according to Mathieu Morlighem),
so it is defensible to raise the topography to better match the observed ice flow.

With more code mods, it would be possible to adjust the input topography in multiple regions.
For now, though, the option is being used only for a single Thwaites region.

This commit is BFB except when the new option is turned on.