Removed the L1L2 velocity solver
This commit removes the option which_ho_approx = 3, the L1L2 solver.
I was never able to obtain stable results with this solver at reasonable time steps,
and it was long since superseded by the DIVA solver, which is similar in concept
but more stable.

Removing L1L2 is a preparatory step toward building a new DIVA solver that solves for
velocity components on cell edges instead of at cell corners.

This commit is BFB for any of the other velocity solvers.