Rewrote the marine_connection_mask subroutine
Several subroutines use marine_connection_mask, a mask set to 1 for marine-based cells
with a path through other marine-based cells to the ocean.
This commit cleans up and slightly changes subroutine glissade_marine_connection_mask.
The input arguments are now thck, topg, eus, and thklim, which are used to identify
ocean cells that can seed the fill.  A local parameter, ocean_topg_threshold = -500 m,
is defined so that shallow ice-free ocean cells (topg > -500 m) do not seed the fill.
Ocean cells with non-ocean neighbors are also prevented from seeding the fill,
to reduce the chance that landlocked ice-free cells with deep topography (e.g.,
in narrow fjords) will be identified as marine-connected.

The mask depends in general on topg and eus but not thck.
It is now computed at initialization, and recomputed after isostasy updates (if isostasy
is turned on), but no longer is computed every time step by subroutines that use the mask.
It would need to be recomputed if eus were to change during the run.

For bmlt_float_inversion, marine_connection_mask replaces the old bmlt_float_inversion_mask.

I also modified several iterative loops with recursive calls to glissade_fill:
in glissade_marine_connection_mask, glissade_ice_sheet_mask, glissade_lake_mask, and
remove_icebergs.  At the end of each iteration, there is now a global sum to check for convergence,
and the loop is exited once the count has converged.

This commit is BFB in most cases.  I found roundoff-level changes starting around year 3 for a run
with eigencalving, because of a small difference in the marine mask computation.

Exit glissade_fill loops when the count converges