New f_flotation option using an stdev correction of topograpy
This commit adds a new which_ho_flotation_function option: HO_FLOTATION_FUNCTION_LINEAR_STDEV = 4.
The new option has the same logic as the current option 3, except that
when computing the flotation function, topg is replaced with (topg + factor*topg_stdev),
where topg_stdev is the standard deviation of topg in each grid cell,
as computed from a high-resolution grid file.

I set factor = 1, meaning that topg is raised by one standard deviation for the f_flotation_calculation.
In principle, this factor is tunable, but for now it is hardwired to 1.

I computed topg_stdev and added it to the input file as follows:
- I ran a diagnostic-only run (tend = 0) with the 1-km input file.
  After reading the input file, I computed the standard deviation of topg
  in each 8x8 block of the grid and wrote the result to an output file.
- The resulting topg_stdev field on the 1-km grid is highly redundant;
  each cell in an 8x8 block has the same value.  Using NCO, I coarsened the 1-km output
  to 8 km by removing 7 out of 8 values in both the x1 and y1 directions.
- I appended the resulting topg_stdev field to the 8-km input file.
- I repeated this process for the 4-km and 2-km grids, using the 1-km file
  to compute topg_stdev in blocks of size 4x4 and 2x2, respectively.

In the future, it would be cleaner to compute topg_stdev when creating the initial input file.

I added topg_stdev as an I/O field.  It is loadable when present in the input file,
but is used only for the new flotation_function options.
I also added topg_stdev to the restart file for the new option.

The goal of this new option is to increase basal friction in regions of
rough ocean topography, where the topography is under-resolved on the input grid
and is likely missing pinning points.

The correction of f_flotation is similar in spirit to a method of Pollard & DcConto (2012),
who set f_ground ~ 0.5 in cells where h_cavity << topg_stdev, going to 0 as h_cavity -> topg_stdev.

In test spin-ups (using the nonlocal-meanAnt and nonlocal-slope-PIGL basal melt schemes),
the differences with this change are modest but generally desirable:
* Reduced magnitude of the dT_basin cooling correction for the ASE (basin 9)
* Less Thwaites GL retreat and thinning for nonlocal-slope-PIGL (which focuces melting near the GL)
* Less central Ronne retreat and thinning for nonlocal-meanAnt
* Smaller thickness biases for the Moscow U./Totten region

This commit is answer-changing only when using the new flotation function.