Added lat and lon as I/O variables
I added variables lat and lon (latitude and longitude) to the model%general data type
and to glide_vars.def, with units degreeN and degreeE.
CISM input files now often include the lat and lon fields.
It can be convenient to read in these fields and write them to output files.

I added a new option called 'read_lat_lon' in glide_types. This option serves two purposes:
(1) If the user sets read_lat_lon = .true. in the config file, but lat and lon fields
    are not present in the input file, the code will abort.
(2) If the user sets read_lat_lon = .true. and these two field are present in the input file
    as required, then they will be written to the restart file and will be available when
    the model restarts.

The default is read_lat_lon = .false.  In this case, lat and lon will be read from the input file,
if present, because they are 'load = 1' fields, but are not written to the restart file.
In this way, the user controls whether lat and lon are in the restart file.
If they are not needed, then it may be preferable to leave them out to reduce storage.

Currently, there are no code options based on the values of latitude and longitude,
but such options could now be added.

In glide_vars.def, I removed many lines of the form 'coordinates: lon lat'.
As far as I know, the coordinates info is no longer used.

This commit is BFB.