Minor calving changes
(1) Reverted to an earlier version of the iceberg removal criterion, where a calving-front cell
    is protected from removal if it has inactive CF edge neighbors and an active diagonal neighbor.
    With thickness-based calving, there are fewer thickness oscillations if corner CF cells
    are allowed to remain and thicken.

(2) Modified the logic for removing cells that are advected beyond the original CF.
    Cells where any ice was present before advection (not just active ice) are now protected.

(3) Modified the logic for computing a calving thickness threshold, for calving option 6.
    Previously, it was not allowed to compute thck_calving_threshold on restart.
    Now, this field can be computed on restart if not read from the input file.
    This makes it easier to compute thck_calving_threshold after a spin-up.

(4) Changed the diagnostics for floating cell transitions, to distinguish between cells flipping
    back and forth between floating and grounded, versus back and forth between floating and open ocean.

This commit is answer-changing for runs with iceberg removal and a subgrid calving front.