Added an apply_calving_mask option, removed the redundant no_advance option
This commit adds a logical option, apply_calving_mask, which is false by default.
When set to true, it has the same function as marine_marine = CALVING_GRID_MASK = 5;
any ice in grid cells with calving_mask = 1 is removed and added to the calving flux.

If setting apply_calving_mask = .true., the user should set marine_marine = CALVING_NONE = 0,
unless it is desired to apply a calving law in addition to applying a calving mask.

This new option allows the user to apply a no-advance calving mask at the same time
as a calving law (e.g., a thickness threshold, marine_margin = 6).
For now, the redundant option marine_margin = 5 is still supported, so as not to break
existing config files.  Later, option 5 can be removed or repurposed.

I also removed the no_advance field, which served the same purpose as calving_mask.
Originally, no_advance was used to prevent advance on land, and calving_mask to prevent
advance of the floating calving front, but the logic was identical.
Later, calving_mask could be renamed as no_advance_mask, which is a more general description.