Cleaned up restart variables for inversion options
In subroutine define_glide_restart_variables, I modified the logic to better separate the
'compute' and 'apply' options for Cp inversion, bmlt inversion, and bmlt_basin inversion.

In particular, floating_thck_target is now allocated, read in, and written to restart
for HO_BMLT_BASIN_INVERSION_COMPUTE, but not for HO_BMLT_BASIN_INVERSION_APPLY.
This fixes a model failure that occurred when CISM tried to write floating_thck_target
to the restart file without first allocating it.