Support repeat cycling and a time offset in forcing files
For ISMIP6-based Antarctic projections, ocean forcing data is provided only through year 2100.
To continue a run past 2100, we typically cycle repeatedly through the data from years 2081-2100.

Previously, we created forcing files in which the time slices for years 2101-2500 were copies
of the slices from years 2081-2500, but this leads to unnecessarily large files.
With this commit, the user can specify config variables in the [CF forcing] section,
instructing CISM to cycle repeatedly through a selected period of data.

For the example above, we would set
   time_cycle_start = 2081
   nyear_cycle = 20
in the config file.
Then for all years from 2081 through the end of the run, CISM will cycle through the 2081-2100 forcing.

To implement these changes, I added time_cycle_start and nyear_cycle to the glimmer_nc_input derived type,
and added code in subroutine handle_input_forcing to read the new parameters.

I also added a parameter called time_offset to the glimmer_nc_input derived type.
This is the difference between the time variable in the forcing file and the CISM model time.
For example, forcing files can have time = 1995 to denote data from calendar year 1995.
In CISM, t = 1994.0 corresponds to 1 Jan. 1995 (since t = 0.0 corresponds to 1 Jan. of year 1).
To get CISM to apply input forcing from year 1995 during the model time range [1994.0, 1995.0),
the user can now set time_offset = 1 in [CF forcing].
The default is time_offset = 0, which is appropriate for forcing files with a time variable
equal to the time since 1 Jan. 0001 (as opposed to the calendar year).
Thanks to Heiko Goelzer for pointing out the need for this option.

Subroutine NAME_read_forcing in ncdf_template.F90.in has been modified to handle time offsets and cycling.
This subrooutine is used to construct subroutine glide_read_forcing in glide_io.F90, and similarly in
other *io_F90 files.

This commit affects only the reading of forcing files.