Added tridiagonal preconditioners for Blatter-Pattyn
In summer 2019 I added tridiagonal preconditioners (PCs) for the DIVA solver.
These PCs reduce the number of linear iterations per nonlinear iteration
compared to a diagonal PC, resulting in significant code speedup.

There are two versions, local and global.
The global version solves a tridiagonal problem on each row and column of the global domain,
requiring global gathers, whereas the local version solves a tridiagonal problem
on each task and requires no gathers.  Generally, the local version is faster;
the reduced cost per iteration makes up for the less accurate solution.
This might not be the case for compilers or architectures with more efficient gathers.

This commit extends the local and global tridiagonal PCs to the Blatter-Pattyn solver.
This became a priority when I tried to run a BP problem for all of Antarctica
and got low throughput, in part because of poor SIA preconditioning.

The basic strategy is as follows:
(1) Given the BP matrices Auu and Avv, form vertically summed matrices Auu_vsum and Avv_vsum.
    Also form a vertical run of the residual vectors ru and rv.
(2) Pass the summed matrics and vectors to the existing tridiagonal PC (either local or global)
    and return a 2D solution vector.
(3) Pass the full BP matrices to the existing SIA PC, returning a 3D solution vector in each column.
(4) In each column, add the 2D tridiagonal solution to the 3D SIA solution.  Take this as the
    solution z to the preconditioning problem A*z = r.

Thus, the new tridiagonal PCs are actually hybrid tridiagonal-SIA PCs, combining advantages of both.
The SIA PC works well in regions of slow sliding; the tridiagonal PC works well for ice shelves;
and the combination works well when both kinds of physics are present, as in Antarctica.

I tested the new PCs on several problems:
- For the dome test problem, the horizontal-plane stresses are small and the tridiagonal PC
  makes little difference.  The new solvers effectively function as SIA PCs.
- For the shelf test problem, the tridiagonal PC is a major improvement over SIA and diagonal PCs.
- For an all-Antarctic problem, the new tridiagonal PC is also a considerable improvement.

For the Antarctic problem, the global tridiagonal PC is slightly slower than the local version,
so I plan to use the local version for production runs.

For all these problems, the BP solution is close to the DIVA solution, which begs the question
why one would run with BP.  The immediate reason is that a reviewer asked for a comparison
between BP and DIVA for ISMIP6-based Antarctic simulations.  And in general, it's good to have
a reasonably efficient BP solver (e.g., 20x slower than DIVA instead of 100x slower) in cases where
there is any doubt as to the accuracy of DIVA compared to BP.

This commit is BFB except when using the new BP preconditioners.

Note: I am breaking the rule that the code should compile with any new commits.
This version will not quite compile because I changed some interfaces in subroutines called
from glissade_velo_higher.F90.  The interface changes needed for compilation will be included
in the next commit, which includes many additional, unrelated changes in glissade_velo_higher.F90.