Added an option to raise the bed topography in troughs
An earlier commit added the config option 'adjust_input_topography', which could be used
to raise the topography in ocean ridges (i.e., pinning points) while leaving lower
topography unchanged, within a specified rectangular region.

This commit expands that option.  It can now be used to raise the topography in troughs
while leaving higher topography unchanged.

The code decides between these two kinds of adjustment based on the values of the config parameters
topg_max_adjust and topg_no_adjust.  If topg_max_adjust > topg_no_adjust, the high topography is raised.
If topg_max_adjust < topg_no_adjust, the low topography is raised.
(More precisely, the topography is changed by adjust_topg_delta. Positive delta corresponds to raising,
and negative delta to lowering.)

I used this option to raise the topography on the standard 2-km Antarctic grid in the region
with x in the range (513000, 522000) and y in the range (-1659000, -1652000).
In this region there is a low trough which can create a bottleneck and lead to numerical instability.
Raising the trough makes the flow more stable, allowing a longer timestep for ISMIP6-type runs.
For the 2-km grid, the run is stable with 15 steps per year, compared to > 20 steps per year
without the topographic adjustment.

This commit is BFB except when adjust_input_topography = T.